메인 페이지 세부필터/상단 필터 & 변수,함수,모듈명 변경
