+++ controllers/productController.js
@@ -1,17 +1,17 @@
-const { mainService } = require('../services');
+const { productService } = require('../services');
 const { catchAsync } = require('../utils/error');
 
-const requestMain = catchAsync(async ( req, res ) => {
+const getProducts = catchAsync(async ( req, res ) => {
 
 
     const { offset, limit, size, gender, special, price, headerFilter } = req.query;
 
-    const main = await mainService.requestMain(offset, limit, size, gender, special, price, headerFilter );
+    const main = await productService.getProducts(offset, limit, size, gender, special, price, headerFilter );
     res.status(200).json(main);
 });
 
 const requestAllMain = catchAsync(async ( req, res )=> {
-    const main = await mainService.requestAllMain();
+    const main = await productService.requestAllMain();
 
     res.status(201).json(main);
 })
@@ -20,7 +20,7 @@ const categoryFilter = catchAsync( async ( req, res ) => {
 
     const { categoryId } = req.params;
     console.log(categoryId)
-    const categoryFilter = await mainService.categoryFilter(categoryId);
+    const categoryFilter = await productService.categoryFilter(categoryId);
     res.status(201).json(categoryFilter);
 })
-module.exports = { requestMain, requestAllMain,categoryFilter };
+module.exports = { getProducts, requestAllMain,categoryFilter };
