+++ routes/index.js
@@ -1,10 +1,10 @@
 const express = require('express');
-const mainRouter = require('./mainRouter');
+const productRouter = require('./productRouter');
 const userRouter = require('./userRouter'); 
 const router = express.Router();
 
 
-router.use('/products',mainRouter);
+router.use('/products',productRouter);
 router.use('/users', userRouter);
 
 module.exports = router;
